LayoutTests/imported/w3c/web-platform-tests/scroll-animations/animation-trigger/animation-trigger-replay.tentative-expected.txt


FAIL replay behavior restarts a paused animation assert_equals: expected "paused" but got "running"
FAIL replay behavior restarts a finished animation assert_equals: expected "running" but got "paused"
